Bugfix FXIOS-5347 [v109] Search suggestions should be bold (#12532)
* Rename extension

also remove unused method, fix spelling and re-order contents

* Use existing extension for attributed text

* Rename extension file

* Move string extension code and rename file

* Optimize extension
